requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
Pillow==6.1.0
python-dateutil==2.8.0
python-dotenv==0.10.3
oauth2_provider==0.0
django-oauth-toolkit==1.2.0
mysqlclient==1.4.6
django-enumfield==2.0.1

